requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130
torch==2.10.0
biopython==1.86
transformers==5.2.0
accelerate==1.12.0
sympy==1.13.3
notebook==7.5.3
ipykernel==7.2.0
langchain==1.2.10
langgraph==1.0.10
langchain-openai==1.1.10
python-dotenv==1.2.2